refactor: extracted OpenButton to buttons feat: styled userID feat: styled admin index
